requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
mysqlclient>=2.2,<3.0
python-dotenv>=1.0,<2.0

